+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - endpoint support for /external_stores/update_credentials <POST>
     - endpoint support for /hosts/{hostId}/capacity <GET>
     - endpoint support for /hosts/{hostId}/cancel_virtual_controller_shutdown <POST>
+    - endpoint support for /hosts/{hostId}/metrics <GET>
     - endpoint support for /hosts/{hostId}/remove_from_federation <POST>    
     - endpoint support for /hosts/{hostId}/shutdown_virtual_controller <POST>
     - endpoint support for /hosts/{hostId}/hardware <GET>